refactor(crane-context): split sessions.ts DAL into queries/crud/orchestrator
580 LOC mixed-concern file split into three focused modules:
- sessions-queries.ts (193 LOC): reads + staleness helpers + sibling queries
- sessions-crud.ts (289 LOC): writes + heartbeat jitter calculation
- sessions.ts (129 LOC): resumeOrCreateSession orchestrator + barrel re-exports

All consumers unchanged — barrel preserves public API.
