+++ workers/crane-context/src/sessions.ts
@@ -1,389 +1,38 @@
 /**
  * Crane Context Worker - Session Management
  *
- * Core session lifecycle logic: creation, resume, staleness detection, heartbeat.
- * Implements session patterns from ADR 025.
+ * Barrel re-exports + resume-or-create orchestrator.
+ * Read operations: sessions-queries.ts
+ * Write operations: sessions-crud.ts
  */
+
+export {
+  findActiveSessions,
+  getSession,
+  isSessionStale,
+  getStaleThreshold,
+  findSiblingSessions,
+  getSiblingSessionSummaries,
+} from './sessions-queries'
+
+export {
+  createSession,
+  updateHeartbeat,
+  updateSession,
+  endSession,
+  markSessionAbandoned,
+  markSessionsSuperseded,
+  calculateNextHeartbeat,
+} from './sessions-crud'
 
 import type { SessionRecord } from './types'
-import { generateSessionId, nowIso, subtractMinutes, addSeconds } from './utils'
+import { findActiveSessions, getSession, isSessionStale } from './sessions-queries'
 import {
-  STALE_AFTER_MINUTES,
-  HEARTBEAT_INTERVAL_SECONDS,
-  HEARTBEAT_JITTER_SECONDS,
-  CURRENT_SCHEMA_VERSION,
-} from './constants'
-
-// ============================================================================
-// Session Queries
-// ============================================================================
-
-/**
- * Find active session by agent, venture, repo, and track
- * Used for resume logic in POST /sos
- *
- * @param db - D1 database binding
- * @param agent - Agent identifier
- * @param venture - Venture identifier
- * @param repo - Repository (owner/repo format)
- * @param track - Track number (nullable)
- * @returns Array of active sessions (should be 0 or 1, but handles multiple)
- */
-export async function findActiveSessions(
-  db: D1Database,
-  agent: string | null,
-  venture: string | null,
-  repo: string | null,
-  track: number | null
-): Promise<SessionRecord[]> {
-  // All filters are optional. When null, that filter is skipped.
-  // This allows both targeted queries (conflict detection) and
-  // unfiltered queries (remote MCP "show all active sessions").
-  const query = `
-    SELECT * FROM sessions
-    WHERE (? IS NULL OR agent = ?)
-      AND (? IS NULL OR venture = ?)
-      AND (? IS NULL OR repo = ?)
-      AND (? IS NULL OR track = ?)
-      AND status = 'active'
-    ORDER BY last_heartbeat_at DESC
-  `
-
-  const result = await db
-    .prepare(query)
-    .bind(agent, agent, venture, venture, repo, repo, track, track)
-    .all<SessionRecord>()
-
-  return result.results || []
-}
-
-/**
- * Get session by ID
- *
- * @param db - D1 database binding
- * @param sessionId - Session ID
- * @returns Session record or null if not found
- */
-export async function getSession(db: D1Database, sessionId: string): Promise<SessionRecord | null> {
-  const result = await db
-    .prepare('SELECT * FROM sessions WHERE id = ?')
-    .bind(sessionId)
-    .first<SessionRecord>()
-
-  return result
-}
-
-// ============================================================================
-// Staleness Detection
-// ============================================================================
-
-/**
- * Check if a session is stale based on last_heartbeat_at
- * A session is stale if last_heartbeat_at > STALE_AFTER_MINUTES ago
- *
- * @param session - Session record
- * @param staleAfterMinutes - Optional custom staleness threshold (defaults to env var)
- * @returns True if session is stale
- */
-export function isSessionStale(
-  session: SessionRecord,
-  staleAfterMinutes: number = STALE_AFTER_MINUTES
-): boolean {
-  const staleThreshold = subtractMinutes(staleAfterMinutes)
-  return session.last_heartbeat_at < staleThreshold
-}
-
-/**
- * Get staleness threshold timestamp for SQL queries
- * Returns ISO 8601 timestamp representing the staleness cutoff
- *
- * @param staleAfterMinutes - Optional custom staleness threshold
- * @returns ISO 8601 timestamp
- */
-export function getStaleThreshold(staleAfterMinutes: number = STALE_AFTER_MINUTES): string {
-  return subtractMinutes(staleAfterMinutes)
-}
-
-// ============================================================================
-// Session Creation
-// ============================================================================
-
-/**
- * Create a new session
- *
- * @param db - D1 database binding
- * @param params - Session parameters
- * @returns Created session record
- */
-export async function createSession(
-  db: D1Database,
-  params: {
-    agent: string
-    client?: string
-    client_version?: string
-    client_session_id?: string
-    host?: string
-    venture: string
-    repo: string
-    track?: number
-    issue_number?: number
-    branch?: string
-    commit_sha?: string
-    session_group_id?: string
-    actor_key_id: string
-    creation_correlation_id: string
-    meta?: Record<string, unknown>
-  }
-): Promise<SessionRecord> {
-  const now = nowIso()
-  const sessionId = generateSessionId()
-
-  const query = `
-    INSERT INTO sessions (
-      id, agent, client, client_version, client_session_id, host,
-      venture, repo, track, issue_number, branch, commit_sha,
-      status, created_at, started_at, last_heartbeat_at,
-      schema_version, actor_key_id, creation_correlation_id, meta_json,
-      session_group_id
-    ) VALUES (
-      ?, ?, ?, ?, ?, ?,
-      ?, ?, ?, ?, ?, ?,
-      'active', ?, ?, ?,
-      ?, ?, ?, ?,
-      ?
-    )
-  `
-
-  await db
-    .prepare(query)
-    .bind(
-      sessionId,
-      params.agent,
-      params.client || null,
-      params.client_version || null,
-      params.client_session_id || null,
-      params.host || null,
-      params.venture,
-      params.repo,
-      params.track || null,
-      params.issue_number || null,
-      params.branch || null,
-      params.commit_sha || null,
-      now, // created_at
-      now, // started_at
-      now, // last_heartbeat_at
-      CURRENT_SCHEMA_VERSION,
-      params.actor_key_id,
-      params.creation_correlation_id,
-      params.meta ? JSON.stringify(params.meta) : null,
-      params.session_group_id || null
-    )
-    .run()
-
-  // Fetch and return the created session
-  const session = await getSession(db, sessionId)
-  if (!session) {
-    throw new Error('Failed to create session')
-  }
-
-  return session
-}
-
-// ============================================================================
-// Session Updates
-// ============================================================================
-
-/**
- * Update session heartbeat timestamp
- * Used by POST /heartbeat and POST /update
- *
- * @param db - D1 database binding
- * @param sessionId - Session ID
- * @returns Updated timestamp
- */
-export async function updateHeartbeat(
-  db: D1Database,
-  sessionId: string,
-  clientSessionId?: string
-): Promise<string> {
-  const now = nowIso()
-
-  // Backfill client_session_id only when currently NULL — handles in-flight
-  // sessions that started before the column existed. COALESCE keeps any value
-  // already set, including the one captured at /sos.
-  if (clientSessionId) {
-    await db
-      .prepare(
-        'UPDATE sessions SET last_heartbeat_at = ?, client_session_id = COALESCE(client_session_id, ?) WHERE id = ?'
-      )
-      .bind(now, clientSessionId, sessionId)
-      .run()
-  } else {
-    await db
-      .prepare('UPDATE sessions SET last_heartbeat_at = ? WHERE id = ?')
-      .bind(now, sessionId)
-      .run()
-  }
-
-  return now
-}
-
-/**
- * Update session fields (for POST /update)
- *
- * @param db - D1 database binding
- * @param sessionId - Session ID
- * @param updates - Fields to update
- * @returns Updated timestamp
- */
-export async function updateSession(
-  db: D1Database,
-  sessionId: string,
-  updates: {
-    branch?: string
-    commit_sha?: string
-    meta?: Record<string, unknown>
-    client_session_id?: string
-  }
-): Promise<string> {
-  const now = nowIso()
-  const fields: string[] = ['last_heartbeat_at = ?']
-  const bindings: (string | null)[] = [now]
-
-  if (updates.branch !== undefined) {
-    fields.push('branch = ?')
-    bindings.push(updates.branch || null)
-  }
-
-  if (updates.commit_sha !== undefined) {
-    fields.push('commit_sha = ?')
-    bindings.push(updates.commit_sha || null)
-  }
-
-  if (updates.meta !== undefined) {
-    fields.push('meta_json = ?')
-    bindings.push(JSON.stringify(updates.meta))
-  }
-
-  // Backfill client_session_id only when currently NULL — handles in-flight
-  // sessions that started before the column existed.
-  if (updates.client_session_id) {
-    fields.push('client_session_id = COALESCE(client_session_id, ?)')
-    bindings.push(updates.client_session_id)
-  }
-
-  const query = `UPDATE sessions SET ${fields.join(', ')} WHERE id = ?`
-  bindings.push(sessionId)
-
-  await db
-    .prepare(query)
-    .bind(...bindings)
-    .run()
-
-  return now
-}
-
-/**
- * End a session
- * Used by POST /eos
- *
- * @param db - D1 database binding
- * @param sessionId - Session ID
- * @param end_reason - Reason for ending (manual, stale, superseded, error)
- * @returns Ended timestamp
- */
-export async function endSession(
-  db: D1Database,
-  sessionId: string,
-  end_reason: string = 'manual',
-  last_activity_at?: string
-): Promise<string> {
-  const now = nowIso()
-
-  await db
-    .prepare(
-      'UPDATE sessions SET status = ?, ended_at = ?, end_reason = ?, last_activity_at = ? WHERE id = ?'
-    )
-    .bind('ended', now, end_reason, last_activity_at || null, sessionId)
-    .run()
-
-  return now
-}
-
-/**
- * Mark session as abandoned (for stale sessions)
- *
- * @param db - D1 database binding
- * @param sessionId - Session ID
- */
-export async function markSessionAbandoned(db: D1Database, sessionId: string): Promise<void> {
-  // Use last_heartbeat_at as ended_at and last_activity_at for abandoned sessions
-  await db
-    .prepare(
-      `UPDATE sessions
-       SET status = 'abandoned',
-           ended_at = last_heartbeat_at,
-           last_activity_at = last_heartbeat_at,
-           end_reason = 'stale'
-       WHERE id = ?`
-    )
-    .bind(sessionId)
-    .run()
-}
-
-/**
- * Mark multiple sessions as superseded
- * Used when /sod finds multiple active sessions for same tuple
- *
- * @param db - D1 database binding
- * @param sessionIds - Array of session IDs to mark as superseded
- */
-export async function markSessionsSuperseded(db: D1Database, sessionIds: string[]): Promise<void> {
-  if (sessionIds.length === 0) return
-
-  const now = nowIso()
-  const placeholders = sessionIds.map(() => '?').join(',')
-
-  const query = `
-    UPDATE sessions
-    SET status = 'ended',
-        ended_at = ?,
-        end_reason = 'superseded'
-    WHERE id IN (${placeholders})
-  `
-
-  await db
-    .prepare(query)
-    .bind(now, ...sessionIds)
-    .run()
-}
-
-// ============================================================================
-// Heartbeat Jitter Calculation
-// ============================================================================
-
-/**
- * Calculate next heartbeat timestamp with jitter
- * Returns random interval between base ± jitter seconds
- *
- * @returns Object with next_heartbeat_at timestamp and actual interval used
- */
-export function calculateNextHeartbeat(): {
-  next_heartbeat_at: string
-  heartbeat_interval_seconds: number
-} {
-  // Generate random jitter: ±HEARTBEAT_JITTER_SECONDS
-  const jitter =
-    Math.floor(Math.random() * (HEARTBEAT_JITTER_SECONDS * 2 + 1)) - HEARTBEAT_JITTER_SECONDS
-
-  const intervalSeconds = HEARTBEAT_INTERVAL_SECONDS + jitter
-  const nextHeartbeat = addSeconds(intervalSeconds)
-
-  return {
-    next_heartbeat_at: nextHeartbeat,
-    heartbeat_interval_seconds: intervalSeconds,
-  }
-}
+  createSession,
+  updateHeartbeat,
+  markSessionAbandoned,
+  markSessionsSuperseded,
+} from './sessions-crud'
 
 // ============================================================================
 // Session Resume Logic (POST /sos)
@@ -477,104 +126,4 @@ export async function resumeOrCreateSession(
 
   // 6. Create new session (no active session found, or all were stale)
   return await createSession(db, params)
-}
-
-// ============================================================================
-// Session Grouping (Sibling Sessions)
-// ============================================================================
-
-/**
- * Find sibling sessions in the same group
- * Returns active sessions with the same session_group_id
- *
- * @param db - D1 database binding
- * @param sessionGroupId - Group ID to search for
- * @param excludeSessionId - Optional session ID to exclude (current session)
- * @returns Array of sibling session records
- */
-export async function findSiblingSessions(
-  db: D1Database,
-  sessionGroupId: string,
-  excludeSessionId?: string
-): Promise<SessionRecord[]> {
-  let query = `
-    SELECT * FROM sessions
-    WHERE session_group_id = ?
-      AND status = 'active'
-  `
-  const bindings: string[] = [sessionGroupId]
-
-  if (excludeSessionId) {
-    query += ' AND id != ?'
-    bindings.push(excludeSessionId)
-  }
-
-  query += ' ORDER BY last_heartbeat_at DESC'
-
-  const result = await db
-    .prepare(query)
-    .bind(...bindings)
-    .all<SessionRecord>()
-
-  return result.results || []
-}
-
-/**
- * Get session group summary for sibling awareness
- * Returns summary info about sibling sessions without full payload
- *
- * @param db - D1 database binding
- * @param sessionGroupId - Group ID to search for
- * @param excludeSessionId - Optional session ID to exclude
- * @returns Array of sibling session summaries
- */
-export async function getSiblingSessionSummaries(
-  db: D1Database,
-  sessionGroupId: string,
-  excludeSessionId?: string
-): Promise<
-  Array<{
-    id: string
-    agent: string
-    venture: string
-    repo: string
-    track: number | null
-    issue_number: number | null
-    branch: string | null
-    last_heartbeat_at: string
-    created_at: string
-  }>
-> {
-  let query = `
-    SELECT id, agent, venture, repo, track, issue_number, branch,
-           last_heartbeat_at, created_at
-    FROM sessions
-    WHERE session_group_id = ?
-      AND status = 'active'
-  `
-  const bindings: string[] = [sessionGroupId]
-
-  if (excludeSessionId) {
-    query += ' AND id != ?'
-    bindings.push(excludeSessionId)
-  }
-
-  query += ' ORDER BY last_heartbeat_at DESC'
-
-  const result = await db
-    .prepare(query)
-    .bind(...bindings)
-    .all<{
-      id: string
-      agent: string
-      venture: string
-      repo: string
-      track: number | null
-      issue_number: number | null
-      branch: string | null
-      last_heartbeat_at: string
-      created_at: string
-    }>()
-
-  return result.results || []
 }
